Skip newline before python docstring
Docstrings are defined in PEP 257 as "A docstring is a string literal
that occurs as the first statement in a module, function, class, or
method definition." When reuse adds a header, it forcibly adds a
space before the docstring violating this principle.

When the style is set to Python and the first line in the file is a
docstring, do not add a space between the license and the docstring.

Signed-off-by: Charlie Jenkins <charlie@rivosinc.com>